Linear Slide Problem
I'm having an issue in that I can't use a boolean for an if condition. It may be syntax, but it says it's undefined. Other than that, what I'm doing is basically changing stages and each time that happens it changes the position by the specified amount of ticks, either in forward or reverse. Once the error with the boolean is fixed, I can use the position of it at that point relative to it's distance because everytime the motor stops, the position resets to 0, so that we can stop and go back midway through if we want to.